% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csnet.R
\name{csnet}
\alias{csnet}
\title{Construct Cell specific network}
\usage{
csnet(data, c = NULL, alpha = 0.01, boxsize = 0.1, weighted = FALSE)
}
\arguments{
\item{data}{row is gene/species, col is smaple/cell}

\item{c}{numeric, the id of cell you intersested, null means to compute all the sample}

\item{alpha}{numeric, Significant level ;eg. 0.001, 0.01, 0.05 ...}

\item{boxsize}{numeric, default 0.1, to define the upper and lower}

\item{weighted}{False or True}
}
\value{
list, the first is single cell network, the second is the degree matrix
}
\description{
Construct Cell specific network
}
\examples{
num_samples <- 5
num_genes <- 6
gene_matrix <- matrix(runif(num_samples * num_genes), ncol = num_samples)
x <- as.data.frame(gene_matrix )
rownames(x) <- paste0("gene_", 1:num_genes)
colnames(x) <- paste0("sample_", 1:num_samples)
csns <- csnet(x, c = NULL, alpha = 0.01, boxsize = 0.1, weighted = FALSE)
}
\references{
Dai H, Li L, Zeng T, Chen L. Cell-specific network constructed by single-cell RNA sequencing data. Nucleic Acids Res. 2019 Jun 20;47(11):e62. doi: 10.1093/nar/gkz172
}
