% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iENA.R
\name{iENA_sPCC}
\alias{iENA_sPCC}
\title{single sample pearson correlation in iENA}
\usage{
iENA_sPCC(dat, PID = "all", transF = F)
}
\arguments{
\item{dat, }{matrix, row is gene/species, col is smaple/cell}

\item{PID, }{vector, sample id for specific or all}

\item{transF, }{F = net output; T = matrix output}
}
\value{
list or matrix
}
\description{
single sample pearson correlation in iENA
}
\examples{
num_samples <- 5
num_genes <- 6
gene_matrix <- matrix(runif(num_samples * num_genes), ncol = num_samples)
x <- as.data.frame(gene_matrix )
rownames(x) <- paste0("gene_", 1:num_genes)
colnames(x) <- paste0("sample_", 1:num_samples)
iENA_sPCC_res <- iENA_sPCC(t(x))
}
\references{
Yu X, Zhang J, Sun S, Zhou X, Zeng T, Chen L. Individual-specific edge-network analysis for disease prediction. Nucleic Acids Res. 2017 Nov 16;45(20):e170. doi: 10.1093/nar/gkx787 Add to Citavi project by DOI. PMID: 28981699 Add to Citavi project by Pubmed ID; PMCID: PMC5714249.
}
