% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.R
\name{netplot_sin}
\alias{netplot_sin}
\title{network plot for single sample network}
\usage{
netplot_sin(
  matrix,
  layout = "sphere",
  node_size = 3,
  node_color = "orange",
  edge_size = 1,
  edge_color = "gray",
  label_size = 3,
  label_color = "grey50",
  metadata = NULL
)
}
\arguments{
\item{matrix}{sparse adjacent matrix or adjacent matrix including gene id or species name}

\item{layout}{layout of network: one of auto, igraph (from layout_tbl_graph_igraph), dendrogram, manual, linear, matrix, treemap, circlepack, partition or hive. the detail in help of ggraph.}

\item{node_size}{numeric or character}

\item{node_color}{color or one variable from metadata}

\item{edge_size}{numeric}

\item{edge_color}{color or "type"}

\item{label_size}{numeric or "weight_abs"}

\item{label_color}{color or one variable from metadata}

\item{metadata}{metadata information of node}
}
\value{
ggplot class
}
\description{
network plot for single sample network
}
\examples{
num_samples <- 50
num_genes <- 100
gene_matrix <- matrix(runif(num_samples * num_genes), ncol = num_samples)
x <- as.data.frame(gene_matrix )
rownames(x) <- paste0("gene_", 1:num_genes)
colnames(x) <- paste0("sample_", 1:num_samples)
Sweet_res <- Sweet(t(x))
netplot_sin(Sweet_res$sample_1,layout = "circle", node_size = "degree", label_color = "black", edge_color = "type")
}
