% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnb.R
\name{plot_DNB}
\alias{plot_DNB}
\title{Plot DNB}
\usage{
plot_DNB(dnb_df, ...)
}
\arguments{
\item{dnb_df}{a data frame of DNB result.}

\item{...}{additional arguments to ggplot2::ggsave}
}
\description{
Plot DNB attibutes from data frame
}
