% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnb.R
\name{search_dnb}
\alias{search_dnb}
\title{Search Candidate DNB genes}
\usage{
search_dnb(
  sni_obj,
  sid,
  type = "raw",
  min_size = 2,
  max_size = Inf,
  included_genes = NULL,
  all = F,
  verbose = T,
  method = "SSN",
  count = F
)
}
\arguments{
\item{min_size}{minimum gene number in subtree. Default 2.}

\item{max_size}{maximum gene number in subtree. Default Inf.}

\item{included_genes}{genes to be included in subtree. Default NULL.}

\item{all}{should all \@code{included_genes} be in subtree. Default False.}

\item{verbose}{output progress.}

\item{dnb}{a DNB object}

\item{with_ctrl}{if consider control group. Default T.}
}
\value{
a DNB object
}
\description{
Cluster by correlation, caculate DNB scores of all subtrees, and get best DNB
score for each time point
}
\references{
\enumerate{
\item Liu, X., Chang, X., Liu, R., Yu, X., Chen, L., & Aihara, K. (2017). Quantifying critical states of complex diseases using single-sample dynamic network biomarkers. PLoS computational biology, 13(7), e1005633. https://doi.org/10.1371/journal.pcbi.1005633
}
}
