% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{ssn}
\alias{ssn}
\title{SSN:sample-specific networks}
\usage{
ssn(matrix, n_method = .pearsonF, alpha = 0.05, dat_ref = NULL, verbose = F)
}
\arguments{
\item{matrix}{inputdata: row is gene/species, col is sample/cell}

\item{n_method}{function :correlation method for feature-feature including pearson
spearman, kendall, sparcc, mutual information and Context Likelihood}

\item{alpha}{significant cutoff, default =0.05}

\item{dat_ref}{matrix of the reference dataset}

\item{verbose}{logical}
}
\value{
SNI object:
\item{data}{your input table}
\item{data_ref}{your reference data}
\item{SNI_raw}{the raw value of pertubation matrix for each sample}
\item{SNI_z}{the z value of pertubation matrix for each sample}
\item{SNI_adj}{the adjcent matrix or network  for each sample}
\item{SNI_cen}{the degree/betweeness/closeness for each sample}
}
\description{
SSN:sample-specific networks
}
\examples{
num_samples <- 100
num_genes <- 60
gene_matrix <- matrix(runif(num_samples * num_genes), ncol = num_samples)
x <- as.data.frame(gene_matrix )
rownames(x) <- paste0("gene_", 1:num_genes)
colnames(x) <- paste0("sample_", 1:num_samples)
ssn_res <- ssn(t(x))
}
\references{
Liu X, Wang Y, Ji H, Aihara K, Chen L. Personalized characterization of diseases using sample-specific networks. Nucleic Acids Res. 2016 Dec 15;44(22):e164. doi: 10.1093/nar/gkw772
}
\seealso{
Other SNI: 
\code{\link{lioness}()}
}
\concept{SNI}
