% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspg.R
\name{sspg}
\alias{sspg}
\title{Sample-specific perturbation of gene interactions}
\usage{
sspg(dat, normalid = NULL, refnet = NULL)
}
\arguments{
\item{dat}{, matrix row is sample id col is feature}

\item{normalid}{,vector including normal sample id or normal sample index}

\item{refnet, }{background net : a matrix include node1 node2}
}
\value{
list including delta rank matrix and edge perturbation matrix
}
\description{
Sample-specific perturbation of gene interactions
}
\examples{
num_samples <- 5
num_genes <- 6
gene_matrix <- matrix(runif(num_samples * num_genes), ncol = num_samples)
x <- as.data.frame(gene_matrix )
rownames(x) <- paste0("gene_", 1:num_genes)
colnames(x) <- paste0("sample_", 1:num_samples)
sspg_res <- sspg(t(x))
}
\references{
Chen Y, Gu Y, Hu Z, Sun X. Sample-specific perturbation of gene interactions identifies breast cancer subtypes. Brief Bioinform. 2021 Jul 20;22(4):bbaa268. doi: 10.1093/bib/bbaa268. PMID: 33126248; PMCID: PMC8293822.
}
