% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet.R
\name{Sweet}
\alias{Sweet}
\title{Sample-specific weighted correlation network}
\usage{
Sweet(
  dat,
  s_method = .pearsonF,
  n_method = .pearsonF,
  balance = 0.1,
  PID = "all",
  alpha = 0.05,
  out_z = F,
  dat_ref = NULL
)
}
\arguments{
\item{dat, }{row is gene/species, col is sample/cell}

\item{s_method, }{function correlation method for sample-sample}

\item{n_method, }{function correlation method for feature-feature}

\item{balance, }{numeric}

\item{PID, }{sample id for specific or all}

\item{alpha, }{significant cutoff}

\item{out_z, }{logistic, F = output is z-score matrix, T = adjacent matrix}

\item{dat_ref, }{matrix of the reference dataset}
}
\value{
outlist, a list including specific network
}
\description{
Sample-specific weighted correlation network
}
\examples{
num_samples <- 5
num_genes <- 6
gene_matrix <- matrix(runif(num_samples * num_genes), ncol = num_samples)
x <- as.data.frame(gene_matrix )
rownames(x) <- paste0("gene_", 1:num_genes)
colnames(x) <- paste0("sample_", 1:num_samples)
sweet_res <- Sweet(t(x))
}
\references{
Chen HH, Hsueh CW, Lee CH, Hao TY, Tu TY, Chang LY, Lee JC, Lin CY. SWEET: a single-sample network inference method for deciphering individual features in disease. Brief Bioinform. 2023 Mar 19;24(2):bbad032. doi: 10.1093/bib/bbad032 Add to Citavi project by DOI. PMID: 36719112 Add to Citavi project by Pubmed ID; PMCID: PMC10025435.
}
