% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{ssn}
\alias{ssn}
\title{SSN:sample-specific networks}
\usage{
ssn(matrix, n_method = .pearsonF, transF = F)
}
\arguments{
\item{matrix, }{row is gene/species, col is smaple/cell}

\item{n_method, }{function :correlation method for feature-feature}

\item{transF, }{F means list output; T means matrix}
}
\value{
specific sample network
}
\description{
SSN:sample-specific networks
}
\examples{
num_samples <- 5
num_genes <- 6
gene_matrix <- matrix(runif(num_samples * num_genes), ncol = num_samples)
x <- as.data.frame(gene_matrix )
rownames(x) <- paste0("gene_", 1:num_genes)
colnames(x) <- paste0("sample_", 1:num_samples)
ssn <- ssn(t(x))
}
\references{
Liu X, Wang Y, Ji H, Aihara K, Chen L. Personalized characterization of diseases using sample-specific networks. Nucleic Acids Res. 2016 Dec 15;44(22):e164. doi: 10.1093/nar/gkw772 Add to Citavi project by DOI. Epub 2016 Sep 4. PMID: 27596597 Add to Citavi project by Pubmed ID; PMCID: PMC5159538.
}
